DRF is like a maze
